feat(): implemented movie same user delete function and admin all rights delete function

+++ src/movies/mutations/delete-movie.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from 'express';
-import ErrorService, { ServerSetupError } from 'services/error-service';
+import ErrorService, { ForbiddenError, ServerSetupError } from 'services/error-service';
+import UserModel from 'models/user-model';
 import { MovieViewModel } from '../types';
 import MoviesModel from '../model';
 
@@ -13,7 +14,12 @@ export const deleteMovie: RequestHandler<
 
     try {
         if (id === undefined) throw new ServerSetupError();
+        if (req.authData === undefined) throw new ServerSetupError();
+        const user = await UserModel.getUserByEmail(req.authData.email);
         const movie = await MoviesModel.getMovie(id);
+
+        if (user.role !== 'ADMIN' && user.id !== movie.user.id) throw new ForbiddenError();
+
         await MoviesModel.deleteMovie(id);
 
         res.status(200).json(movie);
